% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_covid_data.R
\name{clean_covid_data}
\alias{clean_covid_data}
\title{Clean and select variables from COVID dataset}
\usage{
clean_covid_data(df)
}
\arguments{
\item{df}{A data frame containing the raw COVID dataset}
}
\value{
A cleaned data frame
}
\description{
This function selects relevant columns and removes rows with missing values.
}
