% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{perform backward feature selection on COVID data}
\usage{
feature_selection(df, initial_test_size = 0.3, train_size = 0.7, nvmax = 5)
}
\arguments{
\item{df}{Data frame containing the expected columns}

\item{initial_test_size}{size for initial holdout set}

\item{train_size}{proportion for training after holdout}

\item{nvmax}{maximum number of variables to include}
}
\value{
a list with:
\itemize{
\item train_data: the training set with numeric dates
\item test_data: test set with numeric dates
\item selected_features: character vector of selected features
\item performance: data frame with model metrics
}
}
\description{
perform backward feature selection on COVID data
}
