% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load covid data}
\usage{
load_data(file_path)
}
\arguments{
\item{file_path}{Path to CSV file or URL}
}
\value{
A tibble with the loaded data
}
\description{
Simplified reading
}
