% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_final_lm_model.R
\name{make_final_lm_model}
\alias{make_final_lm_model}
\title{Create results linear regression model with the selected features}
\usage{
make_final_lm_model(df)
}
\arguments{
\item{df}{A data frame with at least \code{date},
\code{new_persons_vaccinated},
\code{new_confirmed},
\code{new_intensive_care_patients},
\code{new_hospitalized_patients}
\code{search_trends_anxiety} columns}
}
\value{
returns the final regression model from our results
}
\description{
Create results linear regression model with the selected features
}
