% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predictions.R
\name{model_predictions}
\alias{model_predictions}
\title{predicts the model's R^2 and RMSE and then saves it to a table}
\usage{
model_predictions(final_model, test_df, output_file)
}
\arguments{
\item{final_model}{A lm model}

\item{test_df}{The test split of the data we will be verifying against. This should have date as a numeric value}

\item{output_file}{Path to save the output}
}
\value{
returns a table with the R^2 and RMSE values
}
\description{
predicts the model's R^2 and RMSE and then saves it to a table
}
