% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_viz.R
\name{plot_anxiety_time_series}
\alias{plot_anxiety_time_series}
\title{Plot anxiety search trends over time}
\usage{
plot_anxiety_time_series(df, output_file)
}
\arguments{
\item{df}{A data frame with at least \code{date} and \code{search_trends_anxiety} columns}

\item{output_file}{Path to save the output figure.}
}
\value{
A ggplot object
}
\description{
Plot anxiety search trends over time
}
