% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_lm_coef.R
\name{results_lm_coef}
\alias{results_lm_coef}
\title{create the regression coefficent table and writes it to a csv in the resulte folder as well.}
\usage{
results_lm_coef(lm_model, output_file)
}
\arguments{
\item{lm_model}{A lm model}

\item{output_file}{Path to save the output}
}
\value{
returns the lm coefficients and a csv with the data saved onto it
}
\description{
create the regression coefficent table and writes it to a csv in the resulte folder as well.
}
